# api-contract/openapi.json
{
  "components" : {
    "schemas" : {
...
      "AdminDepositRequestResponse" : {
        "properties" : {
          "amount" : {
            "type" : "number"
          },
          "cancelled_at" : {
            "format" : "date-time",
            "type" : "string"
          },
          "confirmation_reference" : {
            "type" : "string"
          },
          "confirmed_at" : {
            "format" : "date-time",
            "type" : "string"
          },
          "confirmed_by_user_id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "created_at" : {
            "format" : "date-time",
            "type" : "string"
          },
          "currency_code" : {
            "type" : "string"
          },
          "deposit_method_id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "deposit_method_title" : {
            "type" : "string"
          },
          "details_issued_at" : {
            "format" : "date-time",
            "type" : "string"
          },
          "details_issued_by_user_id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "events" : {
            "items" : {
              "$ref" : "#/components/schemas/MoneyOperationEventResponse"
            },
            "type" : "array"
          },
          "operator_comment" : {
            "type" : "string"
          },
          "payment_details" : {
            "type" : "string"
          },
          "payment_instruction" : {
            "$ref" : "#/components/schemas/DepositPaymentInstructionResponse"
          },
          "public_id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "reject_reason" : {
            "type" : "string"
          },
          "rejected_at" : {
            "format" : "date-time",
            "type" : "string"
          },
          "rejected_by_user_id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "status" : {
            "enum" : [ "PENDING_DETAILS", "WAITING_PAYMENT", "PAYMENT_VERIFICATION", "CONFIRMED", "REJECTED", "EXPIRED", "CANCELLED" ],
            "type" : "string"
          },
          "treasury_account_id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "treasury_transaction_id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "treasury_transactions" : {
            "items" : {
              "$ref" : "#/components/schemas/TreasuryTransactionResponse"
            },
            "type" : "array"
          },
          "updated_at" : {
            "format" : "date-time",
            "type" : "string"
          },
          "user_id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "user_marked_paid_at" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "CreateDepositRequest" : {
        "properties" : {
          "amount" : {
            "minimum" : 1.0E-4,
            "type" : "number"
          },
          "deposit_method_id" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "amount", "deposit_method_id" ],
        "type" : "object"
      },
...
      "DepositPaymentInstructionResponse" : {
        "properties" : {
          "amount" : {
            "type" : "number"
          },
          "currency_code" : {
            "type" : "string"
          },
          "deposit_payment_route_public_id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "expires_at" : {
            "format" : "date-time",
            "type" : "string"
          },
          "issued_at" : {
            "format" : "date-time",
            "type" : "string"
          },
          "issued_by_user_id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "operator_comment" : {
            "type" : "string"
          },
          "payment_details" : {
            "additionalProperties" : { },
            "type" : "object"
          },
          "public_id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "status" : {
            "enum" : [ "ACTIVE", "USED", "CANCELLED" ],
            "type" : "string"
          },
          "treasury_account_code" : {
            "type" : "string"
          },
          "treasury_account_public_id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "treasury_account_title" : {
            "type" : "string"
          },
          "treasury_amount" : {
            "type" : "number"
          },
          "treasury_currency_code" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "DepositRequestResponse" : {
        "properties" : {
          "amount" : {
            "type" : "number"
          },
          "cancelled_at" : {
            "format" : "date-time",
            "type" : "string"
          },
          "confirmed_at" : {
            "format" : "date-time",
            "type" : "string"
          },
          "created_at" : {
            "format" : "date-time",
            "type" : "string"
          },
          "currency_code" : {
            "type" : "string"
          },
          "deposit_method_id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "deposit_method_title" : {
            "type" : "string"
          },
          "details_issued_at" : {
            "format" : "date-time",
            "type" : "string"
          },
          "payment_details" : {
            "type" : "string"
          },
          "payment_instruction" : {
            "$ref" : "#/components/schemas/DepositPaymentInstructionResponse"
          },
          "public_id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "reject_reason" : {
            "type" : "string"
          },
          "rejected_at" : {
            "format" : "date-time",
            "type" : "string"
          },
          "status" : {
            "enum" : [ "PENDING_DETAILS", "WAITING_PAYMENT", "PAYMENT_VERIFICATION", "CONFIRMED", "REJECTED", "EXPIRED", "CANCELLED" ],
            "type" : "string"
          },
          "updated_at" : {
            "format" : "date-time",
            "type" : "string"
          },
          "user_marked_paid_at" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "MoneyOperationEventResponse" : {
        "properties" : {
          "actor_type" : {
            "enum" : [ "USER", "OPERATOR", "SYSTEM" ],
            "type" : "string"
          },
          "actor_user_id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "created_at" : {
            "format" : "date-time",
            "type" : "string"
          },
          "event_type" : {
            "enum" : [ "DEPOSIT_CREATED", "DEPOSIT_DETAILS_ISSUED", "DEPOSIT_USER_MARKED_PAID", "DEPOSIT_CONFIRMED", "DEPOSIT_REJECTED", "DEPOSIT_CANCELLED", "WITHDRAWAL_CREATED", "WITHDRAWAL_CANCELLED", "WITHDRAWAL_TAKEN", "WITHDRAWAL_PAYOUT_PLANNED", "WITHDRAWAL_REJECTED", "WITHDRAWAL_CONFIRMED" ],
            "type" : "string"
          },
          "operation_public_id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "operation_type" : {
            "enum" : [ "DEPOSIT_REQUEST", "WITHDRAWAL_REQUEST" ],
            "type" : "string"
          },
          "operator_note" : {
            "type" : "string"
          },
          "payload" : {
            "additionalProperties" : { },
            "type" : "object"
          },
          "public_id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "public_note" : {
            "type" : "string"
          },
          "status_after" : {
            "type" : "string"
          },
          "status_before" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "PageDepositRequestResponse" : {
        "properties" : {
          "content" : {
            "items" : {
              "$ref" : "#/components/schemas/DepositRequestResponse"
            },
            "type" : "array"
          },
          "empty" : {
            "type" : "boolean"
          },
          "first" : {
            "type" : "boolean"
          },
          "last" : {
            "type" : "boolean"
          },
          "number" : {
            "format" : "int32",
            "type" : "integer"
          },
          "numberOfElements" : {
            "format" : "int32",
            "type" : "integer"
          },
          "pageable" : {
            "$ref" : "#/components/schemas/PageableObject"
          },
          "size" : {
            "format" : "int32",
            "type" : "integer"
          },
          "sort" : {
            "$ref" : "#/components/schemas/SortObject"
          },
          "totalElements" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalPages" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "PageableObject" : {
        "properties" : {
          "offset" : {
            "format" : "int64",
            "type" : "integer"
          },
          "pageNumber" : {
            "format" : "int32",
            "type" : "integer"
          },
          "pageSize" : {
            "format" : "int32",
            "type" : "integer"
          },
          "paged" : {
            "type" : "boolean"
          },
          "sort" : {
            "$ref" : "#/components/schemas/SortObject"
          },
          "unpaged" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "SortObject" : {
        "properties" : {
          "empty" : {
            "type" : "boolean"
          },
          "sorted" : {
            "type" : "boolean"
          },
          "unsorted" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "TreasuryTransactionResponse" : {
        "properties" : {
          "actor_user_id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "amount" : {
            "type" : "number"
          },
          "created_at" : {
            "format" : "date-time",
            "type" : "string"
          },
          "currency_code" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "external_reference" : {
            "type" : "string"
          },
          "group_public_id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "metadata" : {
            "additionalProperties" : { },
            "type" : "object"
          },
          "operation_public_id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "operation_type" : {
            "enum" : [ "DEPOSIT_REQUEST", "WITHDRAWAL_REQUEST" ],
            "type" : "string"
          },
          "operator_comment" : {
            "type" : "string"
          },
          "public_id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "transaction_type" : {
            "enum" : [ "DEPOSIT_IN", "WITHDRAWAL_OUT", "MANUAL_IN", "MANUAL_OUT", "ADJUSTMENT", "TRANSFER_IN", "TRANSFER_OUT" ],
            "type" : "string"
          },
          "treasury_account_code" : {
            "type" : "string"
          },
          "treasury_account_public_id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "treasury_account_title" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "Generated API contract for frontend/backend synchronization.",
    "title" : "Prime Market Backend API",
    "version" : "0.0.1-SNAPSHOT"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/admin/deposit-requests/{publicId}" : {
      "get" : {
        "operationId" : "get_4",
        "parameters" : [ {
          "in" : "path",
          "name" : "publicId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminDepositRequestResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "deposit-request-admin-controller" ]
      }
    },
...
    "/api/backoffice/deposit-requests/{publicId}" : {
      "get" : {
        "operationId" : "get_3",
        "parameters" : [ {
          "in" : "path",
          "name" : "publicId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminDepositRequestResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "deposit-request-admin-controller" ]
      }
    },
...
    "/api/deposit-requests" : {
      "get" : {
        "operationId" : "list_2",
        "parameters" : [ {
          "description" : "Canonical single-value request status. Omit the parameter to return requests of all statuses.",
          "in" : "query",
          "name" : "status",
          "required" : false,
          "schema" : {
            "enum" : [ "PENDING_DETAILS", "WAITING_PAYMENT", "PAYMENT_VERIFICATION", "CONFIRMED", "REJECTED", "EXPIRED", "CANCELLED" ],
            "type" : "string"
          }
        }, {
          "description" : "Zero-based page index (0..N)",
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "default" : 0,
            "minimum" : 0,
            "type" : "integer"
          }
        }, {
          "description" : "The size of the page to be returned",
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "default" : 20,
            "minimum" : 1,
            "type" : "integer"
          }
        }, {
          "description" : "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
          "in" : "query",
          "name" : "sort",
          "required" : false,
          "schema" : {
            "default" : [ "createdAt,DESC" ],
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageDepositRequestResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "deposit-request-controller" ]
      },
      "post" : {
        "operationId" : "create_3",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateDepositRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DepositRequestResponse"
                }
              }
            },
            "description" : "Created"
          }
        },
        "tags" : [ "deposit-request-controller" ]
      }
    },
    "/api/deposit-requests/{publicId}" : {
      "get" : {
        "operationId" : "get_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "publicId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DepositRequestResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "deposit-request-controller" ]
      }
    },
    "/api/deposit-requests/{publicId}/cancel" : {
      "post" : {
        "operationId" : "cancel_2",
        "parameters" : [ {
          "in" : "path",
          "name" : "publicId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DepositRequestResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "deposit-request-controller" ]
      }
    },
    "/api/deposit-requests/{publicId}/mark-paid" : {
      "post" : {
        "operationId" : "markPaid",
        "parameters" : [ {
          "in" : "path",
          "name" : "publicId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DepositRequestResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "deposit-request-controller" ]
      }
    },
...
  },
  "servers" : [ {
    "description" : "Relative to the current backend base URL",
    "url" : "/"
  } ]
}